requirements.txt
ruamel.yaml
sendgrid